+++ spec/autoloading_spec.rb
@@ -4,7 +4,7 @@ describe LHS, type: :request do
   context 'autoloading' do
     it "pre/re-loads all LHS classes initialy,|
         because it's necessary for endpoint-to-record-class-discovery",
-    cleanup_before: false do
+    reset_before: false do
       all_endpoints = LHS::Record::Endpoints.all
       expect(all_endpoints['http://datastore/v2/users']).to be_present
       expect(all_endpoints['http://datastore/v2/users/{id}']).to be_present
